requirements.txt
python-dotenv
flask
flask-wtf
bootstrap-flask
beautifulsoup4
requests
fake-useragent
gunicorn
